+++ src/main/java/com/sysunite/rws/deflecties/PeakReadings.java
@@ -1,5 +1,6 @@
 package com.sysunite.rws.deflecties;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Vector;
 
@@ -8,6 +9,6 @@ import java.util.Vector;
  */
 public class PeakReadings {
   public Integer sequenceId;
-  public Double peakLoad;
-  public List<Double> deflections = new Vector<>();
+  public BigDecimal peakLoad;
+  public List<BigDecimal> deflections = new Vector<>();
 }
